Started work on commandline unit tests
Due to the work on the first unit tests we have to refactor lots of code
mainly situated main.cpp. None of the command line option parsing was
safe and none of the exceptions that are possibly trown were catched.
Work on this issue just began and there is much more to do.