Use patterns for the filename
We have now defined some printf like items that oyu can use to define
the ouput file names. This could be quite handy for you. I have also
added a new test case for this feature, to make sure it works.

In the notebook folder is a new python script that shows how you can
generate a few hundred frames for a zoom in a mandelbrot set video.