New printer class for std::cout wrapping
I have added a new Printer class that overrides the stream operator. I
pass around a shared pointer to an object of this class to use a
std::cout wrapper. This is an elegant solution to add a quiet command
line parameter that will suppress writing to stdout. This does not
affect stderr though.

I still had the cmdparser header file in the source directory although I
never made use of it.